Fix parameter name code completion
Parameter names for PowerShell commands would not show any results.
This was because the display text returned from the language service
does not include the '-' at the start of the parameter name and
the code completion list would not find any matches. Now the
completion list shows the text that will be inserted instead of the
label returned from the language service. This fixes the code
completion for parameters.