Replace Scopes parent node with scopes in Locals window
A Scopes node is added to the Locals window whilst the scope
information is being requested from the PowerShell debug service.
Once this information is returned the Scopes node is replaced with
the scopes so they are displayed directly in the Locals window.
Previously the scopes would have a parent Scopes node which would
need to be expanded.