Remove Start Debugging menu from text editor context menu
Debugging a PowerShell script can be initiated from the Run menu
or the main toolbar run button. There is no need for a separate
menu in the text editor for this. The text editor context menu
still has a Stop menu to stop the current debugger session.
Currently this menu is not enabled when a PowerShell file is
being debugged. Neither does the main toolbar run botton change
to a stop button currently.