Fix order of items in Locals window
Work around ObjectValueTreeView bug which causes an array returned
for an evaluating group ObjectValue to be updated with the items
in the wrong order. The ObjectValueTreeView replaces the existing
displayed item with the first item in the array and then inserts
the other items one at a time directly after the first item in the
tree view which results all items being added in the reverse order
apart from the first one. To workaround this the array items are
reversed, apart from the first one, before notifying the Locals
window that the values have changed.