Fix debugging with MonoDevelop 6.0
The debugger was originally tested with MonoDevelop 6.3 from master
and was not working with MonoDevelop 6.0.

The breakpoints were not being hit. With MonoDevelop 6.0 the
TargetEventArgs needs to have the thread backtrace set when the
debugger stops.

The debugger after hitting a breakpoint and then being execution
continuing so the PowerShell script has been run could not then be
used to debug the script again. With MonoDevelop 6.0 the OnExit method
for the DebuggerSession is not called when the script finishes but
only when it is stopped by the user using the Stop menu.