Fix null reference for stack frame method name.
When the stack frame information is not returned from the debugger in
time then the information from the stop event is used. The source name
in this stop event can be null and resulted in the source location's
method name being null which results in a null reference in
MonoDevelop when displaying the callstack.