Fix start without debugging menu being disabled
Add command handlers to the text editor extension so that the
Project.Run command can be enabled or disabled in the Run menu. This
fixes the start without debugger menu being disabled.

Added a custom command factory and execution command to fix running
PowerShell scripts without the debugger on Windows. Running the
PowerShell script file directly fails on Windows since the PowerShell
executable is not associated with these files. So a custom execution
command is used so the full path to the PowerShell executable can be
set and also the execution policy is set to unrestricted. This allows
start without debugging to work on Windows.