Fix code completion not working in new PowerShell file created
After creating a new PowerShell file the text editor would not attach
the PowerShell text editor extension. The text editor extension
lists the PowerShell file extensions as supported in the .addin.xml
file but the filename passed to the text editor for a new file is
null so the text editor is not attached. As a workaround the
PowerShell text editor extension does not specify the file extensions
in the .addin.xml file so it is always attached to every text editor
document. To restrict the text editor extension from being enabled
for every file the IsValidInContext method in the text editor
extension is overridden and a check is made that the file is a
PowerShell file. This prevents the text editor extension from being
initialized for non-PowerShell files but allows it to support new
PowerShell files when they are created from within MonoDevelop.