Show debugger expression tooltip on hovering the mouse when debugging
Tooltips are only shown when debugging. To get this to work a dummy
PowerShell parser has been added which does nothing part from return
a non-null ParsedDocument. This is needed for the default debugger
tooltip provider to work. Also implemented the
IDebuggerExpressionResolver interface on the text editor extension
so the word under the caret is resolved. Without this the '$' at the
start of any variables is not used as part of the expression passed
to the PowerShell debug client.