Show PowerShell documentation for functions
When a completion item is selected a separate tooltip is displayed for
that item with a description. The language service provides a
completion resolve method which can be used to populate PowerShell
documentation for PowerShell functions. This is now called when
the tooltip is requested so PowerShell documentation can be displayed
if installed locally.