Improve function code completion tooltip
When PowerShell has documentation the text description of the
PowerShell function is shown first followed by the detailed
description which has the function overloads with parameters. The
CodeCompletionItem.Description has the short description of the
function after it has been resolved if the documentation is installed
locally, which should be the case for PowerShell Core. The
CompletionItem.Detail has the function overloads and parameters.