Add workaround for debugger tooltips showing ... on Mac
It seems that on the Mac the debugger tooltip window does not resize
enough to fit the text returned after evaluating a variable. The
tooltip window shows '...' or a part of the text followed by '...' so
as a workaround the evaluation is run on a separate task which is then
has Wait called using the EvaluationOptions.EvaluationTimeOut. The
value will usually be evaluated before this timeout occurs so which
results in the debugger tooltip showing the full text. On Windows
this problem does not occur.