Fix error getting tooltip for new unsaved file
The tooltip provider was using the Editor's FileName which is not set
for unsaved files. The DocumentContext.Name is set so that is now
used instead. This prevents a new PowerShell session from being
created for the file and prevents an error being reported about
being unable to start the editor services.