Allow parameters to be passed when debugging PowerShell script
When selecting Run - Debug Application the parameters specified to
a PowerShell script are used when debugging the script. Previously
the parameters were not passed. The working directory and environment
variables are also used when debugging the PowerShell script using
the Debug Application menu.

Using the Debug Application menu is fine for a one off test of some
arguments but the path to the PowerShell script has to be specified
every time. The dialog that is displayed is an internal class used
by MonoDevelop. Ideally there would be a way to pre-popuate certain
values such as the script filename and also remember the information
used between debug runs.