Show launch configurations
Added an Active Configuration menu to the Run menu. This menu can
be expanded and will show configurations read from the launch.json
file. Only PowerShell configurations are shown that are launch
configurations and have a script file defined. The Active Configuration
menu is not shown if a launch.json file cannot be found. The locations
checked are in the .vscode directory relative to the PowerShell file
open in the text editor, and then in the same directory as the
PowerShell file. The Active Configuration by default shows None selected
which means that no configuration will be used. The other configurations
can be selected but currently they are not used at run or debug time.