Fix IDE crash when PowerShell script is edited when debugging
Editing a PowerShell script file when debugging and causing a
breakpoint to move will trigger the debugger session's
OnUpdateBreakEvent method to be called. This was throwing a not
implemented exception causing an IDE crash. Moving breakpoints to
different lines is not supported. Instead of throwing an exception
this method now does nothing.