Fix breakpoints not hit in loops.
With a breakpoint inside a for loop the breakpoint would not be hit.
The problem seems to be that sending a column value for the breakpoint
prevents the PowerShell debugger from hitting it. The PowerShell
extension for Visual Studio Code does not send a column value and
the addin now does the same which fixes the breakpoint not being hit.