Add partial support for breakpoint hit counts
The PowerShell Editor Services debugger supports breakpoint hit counts
that exactly match some number. The Edit Breakpoints dialog has more
hit count options which are not directly supported. An attempt is
made to try to match some of these options but in general they will
not work as expected. There is no way to override the information
supplied in the Edit Breakpoints dialog. The only real solution would
be to create a different dialog.