Fix failure to launch editor services on Mac
Another space in the path problem. However this seems to be specific
to PowerShell when not running on Windows. The editor services is
run with a command line similar to:

powershell
-NoProfile
-NonInteractive
-Command "& '/path/to/addin/scripts/Start-EditorServices.ps1'"
-EditorServicesVersion '0.9.0'
-HostName 'MonoDevelopHost'
-HostProfileId 'MonoDevelop'
-HostVersion '0.1'
-BundledModulesPath '/path/to/addin/modules'
-LogLevel 'Normal'
-LogPath '/Users/user/Library/Logs/XamarinStudio-6.0/powershell-output.log'

The LogPath should not have spaces on any platform. However the
BundledModulesPath on the Mac will contain a space if the addin is
installed. It will be something similar to:

/Users/user/Library/Application Support/XamarinStudio-6.0/LocalInstall/powershelladdin/modules

Now even if this path is quoted with single quotes the PowerShell
editor services cannot be run. You see an error similar to:

A positional parameter cannot be found that accepts argument
'Support/XamarinStudio-6.0/LocalInstall/powershelladdin/modules'

Which seems to be as though PowerShell is removing the single quotes
completely before binding the parameters to those defined in the
script.

As a workaround on the Mac the -BundledModulesPath is not specified
and instead the Start-EditorServices.ps1 script locates the modules
directly using a relative path based on its location.