Fi startup error on Mac with PowerShell 6 Alpha 18
The PowerShellEditorServices module fails to be imported when
PowerShell 6 Alpha 18 is installed. Previous versions of PowerShell
do not have this problem. The PSModulePath casing has been changed
from PSMODULEPATH to PSModulePath in PowerShell 6 Alpha 18. The
startup script has been changed to use the new case for PSModulePath.
Older versions of PowerShell 6 will not work with this change.
 Please enter the commit message for your changes. Lines starting