Support MonoDevelop 8.1
Having to use the RoslynDocumentContext since using a custom
document controller extension and custom document context does not
allow the default debug tooltip provider to show the variable
values when debugging. The debugger tool provider does not work
if the Document.DocumentContext is null and this is only set if
the document has an associated RoslynDocumentExtension.